feat(client): integrated gameApp inside the main app
